Namespace refactoring.
