+++ src/test/java/cloud/filibuster/junit/tests/filibuster/full/local_process/exhaustive/JUnitFilibusterTestWithExhaustiveAnalysisFileByAnnotation.java
@@ -1,9 +1,10 @@
-package cloud.filibuster.junit.tests.filibuster.local_process.extended;
+package cloud.filibuster.junit.tests.filibuster.full.local_process.exhaustive;
 
 import cloud.filibuster.examples.Hello;
 import cloud.filibuster.examples.HelloServiceGrpc;
 import cloud.filibuster.instrumentation.helpers.Networking;
 import cloud.filibuster.junit.FilibusterTest;
+import cloud.filibuster.junit.configuration.FilibusterGrpcExhaustiveAnalysisConfigurationFile;
 import cloud.filibuster.junit.interceptors.GitHubActionsSkipInvocationInterceptor;
 import cloud.filibuster.junit.server.backends.FilibusterLocalProcessServerBackend;
 import cloud.filibuster.junit.tests.filibuster.JUnitBaseTest;
@@ -18,6 +19,7 @@ import org.junit.jupiter.api.extension.ExtendWith;
 
 import java.util.ArrayList;
 import java.util.List;
+
 import java.util.concurrent.TimeUnit;
 
 import static cloud.filibuster.junit.Assertions.wasFaultInjected;
@@ -25,21 +27,33 @@ import static org.junit.jupiter.api.Assertions.assertEquals;
 
 @TestMethodOrder(MethodOrderer.OrderAnnotation.class)
 @SuppressWarnings("Java8ApiChecker")
-public class JUnitFilibusterTestWithBasicAnalysisFileByAnnotation extends JUnitBaseTest {
-    private static final List<String> basicGrpcErrorCodeList = new ArrayList<>();
+public class JUnitFilibusterTestWithExhaustiveAnalysisFileByAnnotation extends JUnitBaseTest {
+    private static final List<String> exhaustiveGrpcErrorCodeList = new ArrayList<>();
 
     static {
-        basicGrpcErrorCodeList.add("DEADLINE_EXCEEDED");
-        basicGrpcErrorCodeList.add("UNAVAILABLE");
-        basicGrpcErrorCodeList.add("INTERNAL");
-        basicGrpcErrorCodeList.add("UNIMPLEMENTED");
+        exhaustiveGrpcErrorCodeList.add("CANCELLED");
+        exhaustiveGrpcErrorCodeList.add("UNKNOWN");
+        exhaustiveGrpcErrorCodeList.add("INVALID_ARGUMENT");
+        exhaustiveGrpcErrorCodeList.add("DEADLINE_EXCEEDED");
+        exhaustiveGrpcErrorCodeList.add("NOT_FOUND");
+        exhaustiveGrpcErrorCodeList.add("ALREADY_EXISTS");
+        exhaustiveGrpcErrorCodeList.add("PERMISSION_DENIED");
+        exhaustiveGrpcErrorCodeList.add("RESOURCE_EXHAUSTED");
+        exhaustiveGrpcErrorCodeList.add("FAILED_PRECONDITION");
+        exhaustiveGrpcErrorCodeList.add("ABORTED");
+        exhaustiveGrpcErrorCodeList.add("OUT_OF_RANGE");
+        exhaustiveGrpcErrorCodeList.add("UNIMPLEMENTED");
+        exhaustiveGrpcErrorCodeList.add("INTERNAL");
+        exhaustiveGrpcErrorCodeList.add("UNAVAILABLE");
+        exhaustiveGrpcErrorCodeList.add("DATA_LOSS");
+        exhaustiveGrpcErrorCodeList.add("UNAUTHENTICATED");
     }
 
     private static int numberOfTestsExceptionsThrownFaultsInjected = 0;
 
     @DisplayName("Test partial hello server grpc route with Filibuster. (MyHelloService, MyWorldService)")
     @ExtendWith(GitHubActionsSkipInvocationInterceptor.class)
-    @FilibusterTest(serverBackend=FilibusterLocalProcessServerBackend.class)
+    @FilibusterTest(analysisConfigurationFile=FilibusterGrpcExhaustiveAnalysisConfigurationFile.class, serverBackend=FilibusterLocalProcessServerBackend.class)
     @Order(1)
     public void testMyHelloAndMyWorldServiceWithFilibuster() throws InterruptedException {
         ManagedChannel helloChannel = ManagedChannelBuilder
@@ -58,7 +72,7 @@ public class JUnitFilibusterTestWithBasicAnalysisFileByAnnotation extends JUnitB
 
                 boolean found = false;
 
-                for (String errorCode: basicGrpcErrorCodeList) {
+                for (String errorCode: exhaustiveGrpcErrorCodeList) {
                     String expectedString = "DATA_LOSS: io.grpc.StatusRuntimeException: " + errorCode;
                     if(t.getMessage().equals(expectedString)) {
                         found = true;
@@ -82,6 +96,6 @@ public class JUnitFilibusterTestWithBasicAnalysisFileByAnnotation extends JUnitB
     @Test
     @Order(2)
     public void testNumAssertions() {
-        assertEquals(4, numberOfTestsExceptionsThrownFaultsInjected);
+        assertEquals(16, numberOfTestsExceptionsThrownFaultsInjected);
     }
 }
