+++ src/test/java/cloud/filibuster/junit/tests/filibuster/full/local_process/exhaustive/JUnitFilibusterTestWithExhaustiveAnalysisFileMaxIterationsNotEnough.java
@@ -1,4 +1,4 @@
-package cloud.filibuster.junit.tests.filibuster.local_process.custom;
+package cloud.filibuster.junit.tests.filibuster.full.local_process.exhaustive;
 
 import cloud.filibuster.examples.Hello;
 import cloud.filibuster.examples.HelloServiceGrpc;
@@ -18,66 +18,67 @@ import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.TestMethodOrder;
 import org.junit.jupiter.api.extension.ExtendWith;
 
+import java.util.ArrayList;
+import java.util.List;
 import java.util.Map;
 import java.util.concurrent.TimeUnit;
 
 import static cloud.filibuster.junit.Assertions.wasFaultInjected;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 
-/**
- * Verify that the Filibuster analysis can be configured to use a custom set of faults.
- */
 @TestMethodOrder(MethodOrderer.OrderAnnotation.class)
 @SuppressWarnings("Java8ApiChecker")
-public class JUnitFilibusterTestWithCustomAnalysisFile extends JUnitBaseTest {
-    private static final String analysisFilePath = "/tmp/filibuster-custom-analysis-file";
+public class JUnitFilibusterTestWithExhaustiveAnalysisFileMaxIterationsNotEnough extends JUnitBaseTest {
+    private static final String analysisFilePath = "/tmp/filibuster-exhaustive-analysis-file";
+    private static final List<String> exhaustiveGrpcErrorCodeList = new ArrayList<>();
 
     static {
-        @SuppressWarnings("Java8ApiChecker")
-        FilibusterAnalysisConfiguration filibusterAnalysisConfiguration = new FilibusterAnalysisConfiguration.Builder()
+        exhaustiveGrpcErrorCodeList.add("CANCELLED");
+        exhaustiveGrpcErrorCodeList.add("UNKNOWN");
+        exhaustiveGrpcErrorCodeList.add("INVALID_ARGUMENT");
+        exhaustiveGrpcErrorCodeList.add("DEADLINE_EXCEEDED");
+        exhaustiveGrpcErrorCodeList.add("NOT_FOUND");
+        exhaustiveGrpcErrorCodeList.add("ALREADY_EXISTS");
+        exhaustiveGrpcErrorCodeList.add("PERMISSION_DENIED");
+        exhaustiveGrpcErrorCodeList.add("RESOURCE_EXHAUSTED");
+        exhaustiveGrpcErrorCodeList.add("FAILED_PRECONDITION");
+        exhaustiveGrpcErrorCodeList.add("ABORTED");
+        exhaustiveGrpcErrorCodeList.add("OUT_OF_RANGE");
+        exhaustiveGrpcErrorCodeList.add("UNIMPLEMENTED");
+        exhaustiveGrpcErrorCodeList.add("INTERNAL");
+        exhaustiveGrpcErrorCodeList.add("UNAVAILABLE");
+        exhaustiveGrpcErrorCodeList.add("DATA_LOSS");
+        exhaustiveGrpcErrorCodeList.add("UNAUTHENTICATED");
+
+        FilibusterAnalysisConfiguration.Builder filibusterAnalysisConfigurationBuilder = new FilibusterAnalysisConfiguration.Builder()
                 .name("java.grpc")
-                .pattern("(.*Service/.*)")
-                .exception("io.grpc.StatusRuntimeException", Map.of(
-                        "cause", "",
-                        "code", "UNAVAILABLE"
-                ))
-                .exception("io.grpc.StatusRuntimeException", Map.of(
-                        "cause", "",
-                        "code", "DEADLINE_EXCEEDED"
-                ))
-                .exception("io.grpc.StatusRuntimeException", Map.of(
-                        "cause", "",
-                        "code", "INVALID_ARGUMENT"
-                ))
-                .exception("io.grpc.StatusRuntimeException", Map.of(
-                        "cause", "",
-                        "code", "NOT_FOUND"
-                ))
-                .build();
-        FilibusterCustomAnalysisConfigurationFile filibusterAnalysisConfigurationFile = new FilibusterCustomAnalysisConfigurationFile.Builder()
+                .pattern("(.*Service/.*)");
+
+        for (String errorCode: exhaustiveGrpcErrorCodeList) {
+            filibusterAnalysisConfigurationBuilder.exception("io.grpc.StatusRuntimeException", Map.of(
+                    "cause", "",
+                    "code", errorCode
+            ));
+        }
+
+        FilibusterAnalysisConfiguration filibusterAnalysisConfiguration = filibusterAnalysisConfigurationBuilder.build();
+        FilibusterCustomAnalysisConfigurationFile filibusterCustomAnalysisConfigurationFile = new FilibusterCustomAnalysisConfigurationFile.Builder()
                 .analysisConfiguration(filibusterAnalysisConfiguration)
                 .build();
-        filibusterAnalysisConfigurationFile.writeToDisk(analysisFilePath);
+        filibusterCustomAnalysisConfigurationFile.writeToDisk(analysisFilePath);
     }
 
     private static int numberOfTestsExceptionsThrownFaultsInjected = 0;
 
-    /**
-     * Inject faults between Hello and World service and verify that all faults that are injected are supposed to be.
-     *
-     * @throws InterruptedException thrown if the gRPC channel fails to terminate.
-     */
     @DisplayName("Test partial hello server grpc route with Filibuster. (MyHelloService, MyWorldService)")
-    @FilibusterTest(analysisFile=analysisFilePath, serverBackend=FilibusterLocalProcessServerBackend.class)
     @ExtendWith(GitHubActionsSkipInvocationInterceptor.class)
+    @FilibusterTest(analysisFile=analysisFilePath, maxIterations=3, serverBackend=FilibusterLocalProcessServerBackend.class)
     @Order(1)
     public void testMyHelloAndMyWorldServiceWithFilibuster() throws InterruptedException {
         ManagedChannel helloChannel = ManagedChannelBuilder
                 .forAddress(Networking.getHost("hello"), Networking.getPort("hello"))
                 .usePlaintext()
                 .build();
-
-        boolean expected = false;
 
         try {
             HelloServiceGrpc.HelloServiceBlockingStub blockingStub = HelloServiceGrpc.newBlockingStub(helloChannel);
@@ -88,23 +89,16 @@ public class JUnitFilibusterTestWithCustomAnalysisFile extends JUnitBaseTest {
             if (wasFaultInjected()) {
                 numberOfTestsExceptionsThrownFaultsInjected++;
 
-                if (t.getMessage().equals("DATA_LOSS: io.grpc.StatusRuntimeException: NOT_FOUND")) {
-                    expected = true;
+                boolean found = false;
+
+                for (String errorCode: exhaustiveGrpcErrorCodeList) {
+                    String expectedString = "DATA_LOSS: io.grpc.StatusRuntimeException: " + errorCode;
+                    if(t.getMessage().equals(expectedString)) {
+                        found = true;
+                    }
                 }
 
-                if (t.getMessage().equals("DATA_LOSS: io.grpc.StatusRuntimeException: DEADLINE_EXCEEDED")) {
-                    expected = true;
-                }
-
-                if (t.getMessage().equals("DATA_LOSS: io.grpc.StatusRuntimeException: UNAVAILABLE")) {
-                    expected = true;
-                }
-
-                if (t.getMessage().equals("DATA_LOSS: io.grpc.StatusRuntimeException: INVALID_ARGUMENT")) {
-                    expected = true;
-                }
-
-                if (! expected) {
+                if (! found) {
                     throw t;
                 }
             } else {
@@ -116,14 +110,11 @@ public class JUnitFilibusterTestWithCustomAnalysisFile extends JUnitBaseTest {
         helloChannel.awaitTermination(1000, TimeUnit.SECONDS);
     }
 
-    /**
-     * Verify that the correct number of Filibuster tests are generated.
-     */
     @DisplayName("Verify correct number of generated Filibuster tests.")
     @ExtendWith(GitHubActionsSkipInvocationInterceptor.class)
     @Test
     @Order(2)
     public void testNumAssertions() {
-        assertEquals(4, numberOfTestsExceptionsThrownFaultsInjected);
+        assertEquals(2, numberOfTestsExceptionsThrownFaultsInjected);
     }
 }
